requirements.txt
Flask
flask-cors
flasgger

pika

requests

python-telegram-bot

google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2

stripe

supabase

python-dotenv

pytz